[op][n_upsable][ime] Fix predictions for composition
:Release Notes:
Fixed prediction application while the text
composition feature is in progress when using VKB.

:Detailed Notes:
When a prediction is applied
text_model.delete_surrounding_text is sent.
The incorrect text range was used to delete
symbols exchanged by applied prediction.

:Testing Performed:
1. Enable the Korean input language.
2. Launch a WAM-managed web app (e.g., UserGuide).
3. Click on Search input.
4. Switch to Korean language in VKB.
5. Type the text 'ㄱㄱㄱ' using VKB.
6. Move the cursor to the first position.
7. Type the text '신' and '세'
8. Click on an Item (ex.'신세대') among the
   predictions.
9. The prediction text + 'ㄱㄱㄱ'
   (ex.'신세대ㄱㄱㄱ') should be displayed.

:QA Notes:
Test-cases: QWT-376, QWT-397, QWT-413, QWT-424,
            QWT-624, QWT-639, QWT-667, QWT-668,
            QWT-669

Issues Addressed:
[WRN-2634]    Port [op][n_upsable][ime] Fix predictions
              for composition
[PLAT-144992] The selected text + 'ㄱㄱ' displayed
              at the field

Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/315774
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/317621
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>

+++ src/ui/base/ime/linux/neva/input_method_auralinux_neva.cc
@@ -48,14 +48,14 @@ void InputMethodAuraLinuxNeva::OnDeleteRange(int32_t index, uint32_t length) {
           surround_range.set_end(
               std::min(range.GetMax(), composition_range.GetMin()));
         }
-        range.set_start(surround_range.GetMax() + index);
-        range.set_end(surround_range.start() + length);
+        range.set_start(surround_range.start() + index);
+        range.set_end(surround_range.start() + index + length);
       }
     } else {
       res = client->GetTextRange(&range);
     }
 
-    if (res)
+    if (res && range.IsValid())
       client->DeleteRange(range);
   }
 }
